# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: InventorySync API
  description: |
    Enterprise-level inventory management API for Shopify stores at startup-friendly prices.
    
    ## Features
    - 🎯 **Custom Fields**: Unlimited custom fields with JSONB storage
    - 🤖 **Workflow Automation**: Event-driven rules engine  
    - 📊 **Advanced Reporting**: Build reports on any field
    - 🚨 **Smart Alerts**: Template-based alerts with analytics
    - 🔗 **Integrations**: Full REST API for third-party systems
    
    ## Authentication
    All endpoints require either Shopify session authentication or API key authentication via `X-API-Key` header.
  version: 1.0.0
  contact:
    name: InventorySync API Support
    email: api-support@inventorysync.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.inventorysync.com/api
    description: Production server
  - url: https://staging-api.inventorysync.com/api  
    description: Staging server
security:
  - ApiKeyAuth: []
  - ShopifySession: []
paths:
...
  /integrations/{shop_domain}/api-keys:
    post:
      summary: Create API key
      description: Generate a new API key for third-party integrations
      tags:
        - Integrations
      parameters:
        - name: shop_domain
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/APIKeyCreate'
      responses:
        '201':
          description: API key created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_key:
                    type: string
                    example: "isk_AbCdEf123456789"
                  api_secret:
                    type: string
                    example: "secret_XyZ987654321"
                  name:
                    type: string
                  permissions:
                    type: array
                    items:
                      type: string
                  warning:
                    type: string
                    example: "Store this API key securely. It will not be shown again."
  /integrations/{shop_domain}/sync:
    post:
      summary: Sync external data
      description: Sync data from external systems into InventorySync
      tags:
        - Integrations
      security:
        - ApiKeyAuth: []
      parameters:
        - name: shop_domain
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExternalDataSync'
      responses:
        '200':
          description: Data synced successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  entity_type:
                    type: string
                  external_id:
                    type: string
                  synced_at:
                    type: string
                    format: date-time
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    ShopifySession:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://shopify.dev/oauth/authorize
          tokenUrl: https://shopify.dev/oauth/token
          scopes:
            read_products: Read products
            write_products: Write products
            read_inventory: Read inventory
            write_inventory: Write inventory
  schemas:
...
    APIKeyCreate:
      type: object
      required:
        - name
        - permissions
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
        permissions:
          type: array
          minItems: 1
          items:
            type: string
            enum: [inventory:read, inventory:write, products:read, products:write, alerts:read, alerts:write]
        expires_at:
          type: string
          format: date-time
        rate_limit:
          type: integer
          minimum: 10
          maximum: 10000
          default: 1000
    ExternalDataSync:
      type: object
      required:
        - entity_type
        - external_id
        - data
        - source_system
      properties:
        entity_type:
          type: string
          enum: [product, variant, inventory, supplier]
        external_id:
          type: string
        data:
          type: object
        source_system:
          type: string
        sync_timestamp:
          type: string
          format: date-time
...